requirements.txt
python==3.10.14
anndata==0.11.4
celltypist==1.6.3
cellxgene_census==1.16.2
dask==2024.1.1
lightning==2.4.0
matplotlib==3.10.1
merlin_core==23.8.0
merlin_dataloader==23.8.0
numba==0.60.0
numpy==2.2.5
pandas==1.5.3
PyYAML==6.0.2
scanpy==1.11.1
scikit_learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
SPARQLWrapper==2.0.0
torch==2.4.1
torchmetrics==1.4.1
tqdm==4.67.1
